Correct date with muliple components
